+++ TrialSeleniumFW/build.gradle
@@ -25,8 +25,6 @@ dependencies {
 
     implementation group: 'org.testng', name: 'testng', version: '6.14.3'
 
-//    implementation group: 'org.seleniumhq.selenium', name: 'selenium-firefox-driver', version: '3.3.1'
-
     // https://mvnrepository.com/artifact/io.cucumber/gherkin
     implementation group: 'io.cucumber', name: 'gherkin', version: '12.0.0'
 
@@ -38,9 +36,10 @@ dependencies {
 }
 
 test {
+    scanForTestClasses = false
+    def testSuite = System.properties.getProperty("testSuite", 'src/TestRunner.xml')
     useTestNG() {
-        useDefaultListeners = true
-        suites "TestRunner.xml"
-        jvmArgs(["-Ddataproviderthreadcount=1"])
+        suites testSuite
+        jvmArgs(["-Ddataproviderthreadcount=${threads}"])
     }
 }
