+++ TrialSeleniumFW/src/gradle.properties
@@ -1,0 +1,1 @@
+threads = 1
