+++ TrialSeleniumFW/src/main/java/Supports/FrameworkInitiation.java
@@ -1,6 +1,5 @@
 package Supports;
 
-import Reports.Reports;
 import com.aventstack.extentreports.ExtentReports;
 import com.aventstack.extentreports.ExtentTest;
 import com.aventstack.extentreports.reporter.ExtentSparkReporter;
@@ -10,11 +9,8 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.firefox.FirefoxOptions;
-import org.openqa.selenium.remote.DesiredCapabilities;
 
 import java.io.File;
-import java.util.HashMap;
-import java.util.Map;
 
 public class FrameworkInitiation {
