+++ TrialSeleniumFW/src/test/java/Runners/TestRunnerClass.java
@@ -6,14 +6,14 @@ import io.cucumber.testng.TestNGCucumberRunner;
 import org.testng.annotations.DataProvider;
 
 @CucumberOptions(
-        features = {"src/test/resources/"},
+        features = {"src/test/resources/GUITestCases.feature"},
         glue = {"StepDefinitions"}
 )
 
 public class TestRunnerClass extends AbstractTestNGCucumberTests {
     private TestNGCucumberRunner testNGCucumberRunner;
     @Override
-    @DataProvider(parallel = false)
+    @DataProvider(parallel = true)
     public Object[][] scenarios(){
         return super.scenarios();
     }
