+++ TrialSeleniumFW/src/test/resources/APITestCases.feature
@@ -1,7 +1,7 @@
 Feature: Test the search function in OpenWeather page
 
   @API
-  Scenario: TC1_Send request to search weather by city name
+  Scenario: [API-TC1] Send request to search weather by city name
     Given I send request to search weather by city name is "London"
     Then Verify the status code is 200
     And Verify the response body is correct as expected
@@ -12,7 +12,7 @@ Feature: Test the search function in OpenWeather page
 
 
   @API
-  Scenario: TC2_Send request to search weather by zip code
+  Scenario: [API-TC2] Send request to search weather by zip code
     Given I send request to search weather by zip code is "10700,th"
     Then Verify the status code is 200
     And Verify the response body is correct as expected
